Refactor TableOne API integration and remove unused files
Moved TableOne API functions from survivalAnalysisAPI.js to a new tableOneAPI.js for better separation of concerns. Updated asyncThunks.js to use the new API module. Removed unused config and RiskTable component files related to TableOne.
