+++ src/redux/explorer/survivalAnalysisAPI.js
@@ -43,30 +43,3 @@ export function fetchConfig() {
     return { ...data, enabled: isSurvivalAnalysisEnabled(data) };
   });
 }
-
-/** @returns {Promise<ExplorerState['config']['tableOneConfig']>} */
-export function fetchTableOneConfig() {
-  return fetchWithCreds({
-    path: '/analysis/tools/tableone/config',
-    method: 'GET',
-  }).then(({ response, data, status }) => {
-    if (status !== 200) throw response.statusText;
-    return { ...data, enabled: isTableOneEnabled(data) };
-  });
-}
-
-/**
- * @returns {Promise<ExplorerState['tableOneResult']>}
- */
-export function fetchTableOneResult(body) {
-  return fetchWithCreds({
-    path: '/analysis/tools/tableone',
-    method: 'POST',
-    body: JSON.stringify(body),
-  }).then(({ response, data, status }) => {
-    if (status !== 200) {
-      throw response.statusText;
-    }
-    return data;
-  });
-}
